backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
email-validator==2.1.0.post1
flask-jwt-extended==4.5.3
werkzeug==3.0.1
